move loging code to other file
