+++ graph/src/blockchain/block_stream.rs
@@ -1,16 +1,16 @@
 use crate::substreams::Clock;
 use crate::substreams_rpc::response::Message as SubstreamsMessage;
-use crate::substreams_rpc::{BlockScopedData, ModulesProgress};
+use crate::substreams_rpc::BlockScopedData;
 use anyhow::Error;
 use async_stream::stream;
 use futures03::Stream;
-use humantime::format_duration;
 use std::fmt;
 use std::sync::Arc;
 use std::time::Instant;
 use thiserror::Error;
 use tokio::sync::mpsc::{self, Receiver, Sender};
 
+use super::substreams_block_stream::SubstreamsLogData;
 use super::{Block, BlockPtr, Blockchain};
 use crate::anyhow::Result;
 use crate::components::store::{BlockNumber, DeploymentLocator};
@@ -335,75 +335,6 @@ pub trait FirehoseMapper<C: Blockchain>: Send + Sync {
         endpoint: &Arc<FirehoseEndpoint>,
         block: &C::Block,
     ) -> Result<BlockPtr, Error>;
-}
-
-pub struct SubstreamsLogData {
-    pub last_progress: Instant,
-    pub last_seen_block : u64,
-}
-
-impl SubstreamsLogData {
-    pub fn info_string(&self, progress : &ModulesProgress) -> String {
-        format!("Substreams backend graph_out last block is {}, {} stages, {} jobs",
-            self.last_seen_block, progress.stages.len(), progress.running_jobs.len()
-        )
-    }
-    pub fn debug_string(&self, progress : &ModulesProgress) -> String {
-        let len =  progress.stages.len();
-        let mut stages_str = "".to_string();
-        for i in (0..len).rev() {
-            let stage = &progress.stages[i];
-            let range = if stage.completed_ranges.len() > 0 {
-                let b = stage.completed_ranges.iter().map(|x| x.end_block).min();
-                format!(" up to {}", b.unwrap_or(0))
-            } else {
-                "".to_string()
-            };
-            let mlen = stage.modules.len();
-            let module = if mlen == 0 {
-                "".to_string()
-            } else if mlen == 1 {
-                format!(" ({})", stage.modules[0])
-            } else {
-                format!(" ({} +{})", stage.modules[mlen - 1], mlen - 1)
-            };
-            if !stages_str.is_empty() {
-                stages_str.push_str(", ");
-            }
-            stages_str.push_str(&format!("#{}{}{}", i, range, module));
-        }
-        let stage_str = if len > 0 { 
-            format!(" Stages: [{}]", stages_str )
-        } else { 
-            "".to_string()
-        };
-        let mut jobs_str = "".to_string();
-        let jlen = progress.running_jobs.len();
-        for i in 0..jlen {
-            let job = &progress.running_jobs[i];
-            if !jobs_str.is_empty() {
-                jobs_str.push_str(", ");
-            }
-            let duration_str = format_duration(Duration::from_millis(job.duration_ms));
-            jobs_str.push_str(&format!("#{} on Stage {} @ {} | +{}|{} elapsed {}", 
-                                        i, 
-                                        job.stage, 
-                                        job.start_block, 
-                                        job.processed_blocks, 
-                                        job.stop_block - job.start_block,
-                                        duration_str));
-        }
-        let job_str = if jlen > 0 {
-            format!(", Jobs: [{}]", jobs_str)
-        } else {
-            "".to_string()
-        };
-        format!("Substreams backend graph_out last block is {},{}{}", 
-            self.last_seen_block,
-            stage_str,
-            job_str,
-        )
-    }
 }
 
 #[async_trait]
